implement g:lightline.component_raw to specify raw type components. (close #240, #134)
Having raw type configuration in component_type is a kind of
implementation design mistake. We sometimes want to specify both the
type and the raw attribute of a component. The raw component feature
shouldn't be included as one of the types of components. Now we can
specify the raw attribute with component_raw and the type with
component_type.