fix: don't bold last component of deus colorscheme
In the deus colorscheme, the first component is bolded, and the rest of
the components aren't.  However, for replace and visual modes, the last
component is bolded.  This appears to be a copy-paste error because
those incorrect lines are precisely the lines added in commit
b1354c53bd658f91f64a14498d97fb178023f496.  Fix it by making them not
bolded.