Add ability for user-configured advanced mode name
e.g. 'ni' `mode(1)` before falling back to `'n': 'NORMAL'`, etc.

Fallback behaviour starts with the full mode string, and removes a
character until a hit is found in the mode name map. Failing any
advanced mode strings, we end up with `mode(1)[0]`, which is identical
to the previous `mode()` query (single character of basic mode string).

Technically, I implemented a secondary fallback if `mode(1)[0]` fails,
explicitly calling `mode()` instead (legacy behaviour prior to this
commit). This should be theoretically impossible to reach but is kept in
case `mode()` ever changes(?)

The final fallback (both in legacy and here) is that a completely
unimplemented-in-the-map mode string evaluates to ''.

~~Batteries~~ documentation included!